Move first-pass select_rep_comments() into utils. Added test assertion.
